[Refactor] Extract pipeline driver instantiation helper
